# conf/openapi.yaml
openapi: 3.0.0
servers:
  - url: http://localhost:9000
    description: Lithos Client API
info:
  description: |
    API Documentation for the Lithos client. Models used
    here are expected to be consistent across any other implementations of the Lithos client.

  version: 1.0.0
  title: Lithos Client API
  contact:
    url: https://github.com/Lithos-Protocol
...
paths:
...
  /wallet/balances:
    get:
      security:
        - ApiKeyAuth: [api_key]
      summary: Wallet balances and lending status
      description: >
        Spendable ERG, token balances, and — under `collateral` — lender keys, queued and live
        positions, locked block rewards, and the self-collateralization settings in effect.


        `nanoErgs` is the ERG available to spend right now: unreserved wallet boxes plus matured
        block rewards. Reading balances and lending status in one call means both describe the same
        instant.
      operationId: getWalletBalances
      tags:
        - wallet
      responses:
        '200':
          description: Current wallet balances and lending status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WalletBalances'
        '403':
          description: Missing or wrong api_key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '503':
          description: Node or wallet unavailable — it may be locked. Retry.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
...
components:
  schemas:
    ApiError:
      description: Error response from API
      type: object
      required:
        - error
        - reason
        - detail
      properties:
        error:
          type: integer
          description: Error code
          example: 500
        reason:
          type: string
          description: Error message explaining the reason of the error
          example: 'Internal server error'
        detail:
          type: string
          nullable: true
          description: Detailed error description
...
    CollateralLenderKeyStatus:
      description: One wallet address that may hold a lending position
      type: object
      x-tags: collateral-market
      required:
        - address
        - status
      properties:
        address:
          type: string
          example: '3WwbQFz3vgi8ChY9dM1haQLGDeQYe5U9z98bwWW4TqxxEAvyYu5y'
        status:
          type: string
          enum: ['FREE', 'QUEUED', 'ACTIVE', 'RETIRING']
          description: >
            FREE has no position and can join. QUEUED is waiting in the queue. ACTIVE holds a live
            collateral box. RETIRING held one until recently and cannot be reused yet. Only FREE
            keys are available to a join. One key holds at most one position; a second position
            under a key already in the active set is forfeited.
          example: 'ACTIVE'
        position:
          type: integer
          format: int64
          description: Queue position, when QUEUED
          example: 411
    CollateralPositionSummary:
      description: One of this wallet's positions, queued or live
      type: object
      x-tags: collateral-market
      required:
        - kind
        - boxId
        - principalNanoErgs
        - permitLit
        - creationHeight
      properties:
        kind:
          type: string
          enum: ['QUEUED', 'LIVE']
          example: 'QUEUED'
        boxId:
          type: string
          example: 'ce00d6e0f4fd3390c0519bb0d24824e1e6b06978163f4845a7cf2c1c7d8f1b65'
        position:
          type: integer
          format: int64
          description: Queue position, when QUEUED
          example: 411
        principalNanoErgs:
          type: string
          description: Decimal string.
          example: '2915000000'
        permitLit:
          type: string
          description: Permit paid at join. Present for QUEUED positions only.
          example: '19380000000'
        carriedLit:
          type: string
          description: >
            LIT carried on the box — emission plus permit together. Present for LIVE positions only.
            Decimal string.
          example: '19380000000'
        creationHeight:
          type: integer
          format: int32
          example: 484102
    CollateralRewardSummary:
      description: This wallet's locked coinbase rewards
      type: object
      x-tags: collateral-market
      required:
        - boxes
        - unlockedBoxes
        - totalNanoErgs
        - unlockedNanoErgs
      properties:
        boxes:
          type: integer
          format: int32
          description: Reward boxes held, including ones still behind the timelock
          example: 4
        unlockedBoxes:
          type: integer
          format: int32
          description: How many are spendable now
          example: 2
        totalNanoErgs:
          type: string
          description: Every reward box together. Decimal string.
          example: '12000000000'
        unlockedNanoErgs:
          type: string
          description: The part past its timelock. Decimal string.
          example: '6000000000'
        blocksUntilFirstUnlock:
          type: integer
          format: int32
          description: Blocks until the earliest still-locked box matures. Absent when none are locked.
...
    WalletCollateralStatus:
      description: >
        The collateral-market half of the wallet. Lender keys are the expensive part to compute, so
        the client serves them from a short-lived cache rather than rescanning per request.
      type: object
      x-tags: collateral-market
      required:
        - autoCollateralize
        - maxOwnCollateral
        - maxJoinsPerRun
        - maxLenderKeys
        - maxPermitPerJoinLit
        - ownPositionsQueued
        - ownPositionsLive
        - ownBudgetRemaining
        - freeLenderKeys
        - keysDerivable
        - lenderKeys
        - positions
        - rewards
      properties:
        autoCollateralize:
          type: boolean
          description: >
            Whether the client joins the queue on its own timer. **While this is true the client
            refuses manual joins** — the automated pass owns lender-key selection, and a manual join
            can take a key it is already using. Turn it off in the client config to join by hand.
          example: false
        maxOwnCollateral:
          type: integer
          format: int32
          description: Configured ceiling on live positions owned by this wallet
          example: 10
        maxJoinsPerRun:
          type: integer
          format: int32
          description: >
            Most positions one automated pass creates. Does not bound a manual join, which is
            limited by free lender keys, the exposure budget and the wallet balance.
          example: 5
        maxLenderKeys:
          type: integer
          format: int32
          description: Wallet addresses dedicated to lending
          example: 32
        maxPermitPerJoinLit:
          type: string
          description: The permit price above which automated joins stop. Decimal string.
          example: '30000000000000'
        ownPositionsQueued:
          type: integer
          format: int32
          example: 3
        ownPositionsLive:
          type: integer
          format: int32
          example: 7
        ownBudgetRemaining:
          type: integer
          format: int32
          description: How many more positions the configured exposure budget allows
          example: 0
        freeLenderKeys:
          type: integer
          format: int32
          description: >
            Keys with status FREE — positions that can be opened without deriving a new key. A join
            needs one key per position.
          example: 4
        keysDerivable:
          type: integer
          format: int32
          description: >
            Further keys the client may derive before reaching `maxLenderKeys`. A join of N
            positions is possible when `freeLenderKeys + keysDerivable >= N`.
          example: 12
        lenderKeys:
          type: array
          items:
            $ref: '#/components/schemas/CollateralLenderKeyStatus'
        positions:
          type: array
          items:
            $ref: '#/components/schemas/CollateralPositionSummary'
        rewards:
          $ref: '#/components/schemas/CollateralRewardSummary'
...
    WalletToken:
      description: A token balance held by the wallet
      type: object
      x-tags: lithosdex
      required:
        - tokenId
        - amount
      properties:
        tokenId:
          type: string
          example: '86e272f770aa4ef48ad7e6d6dbe98b5069edd8383a403a8edb3ceed8a3d5cea3'
        name:
          type: string
          example: 'LIT'
        decimals:
          type: integer
          format: int32
          example: 4
        amount:
          type: string
          description: Raw token units held. Decimal string.
          example: '5420000000'
    WalletBalances:
      description: >
        This client's wallet, for both LithosDex and the collateral market. One call so a page does
        not have to stitch two together, and so both halves describe the same instant.
      type: object
      x-tags: wallet
      required:
        - primaryAddress
        - nanoErgs
        - tokens
        - collateral
      properties:
        primaryAddress:
          type: string
          description: The wallet's main P2PK address, where change and swept rewards land
          example: '3WwbzR34YmLRy3Z6WLON5hBBkX8Zm1gUHmC7tEXDNRJfHBqhsZr'
        nanoErgs:
          type: string
          description: >
            Unreserved ERG available to spend right now, matured coinbases included — the figure a
            join quote's `affordNow` is decided against. Decimal string.
          example: '182400000000'
        tokens:
          type: array
          description: Every token the wallet holds, including LIT.
          items:
            $ref: '#/components/schemas/WalletToken'
        collateral:
          $ref: '#/components/schemas/WalletCollateralStatus'
...
  securitySchemes:
    ApiKeyAuth:
      description: >
        Api key authorization used by Lithos. The hash of this key must be placed in your config, under `lithos.apiKeyHash`.
        The default value is `hello`.
      type: apiKey
      name: api_key
      in: header